Don't close CURL handler when specified
I think if we pass a CURL handler to the makeRequest method, that method should not close the handler, instead it should relegate that process to the one that open it. This will allow us to reuse the same handler to make multiple calls to the API without keep opening and closing it.